Moved sql based migrations to liquibase xml. Moved ejb transactions to deltaspike (cdi) transactions. Configured jboss datasources with puppet. Introduced repositories and validation for DistributionRules. Added CprDistributionRuleFilter. Added validation for distribution rule filters so only 1 with the same name can be applied to a distribution rule. Created dummy repositories. Got bootstrap data to work, and exposes it on a rest endpoint instead of on application start. Got Unit testing to work with transactions and entitymanager. Wrote a Fabric file for application deployment, see examples.
Liquibase migrations are contained in the application artifact, Liquibase is able to handle this.

+++ src/main/java/dk/os2opgavefordeler/rest/AuthEndpoint.java
@@ -61,7 +61,7 @@ public class AuthEndpoint {
 	@GET
 	@Path("/providers/{providerId}/signin")
 	@NoCache
-	public Response beginAuthentication(@PathParam(value = "providerId") int providerId) {
+	public Response beginAuthentication(@PathParam(value = "providerId") long providerId) {
 		try {
 			final HttpSession session = request.getSession();
 
@@ -94,7 +94,7 @@ public class AuthEndpoint {
 		log.info("finishAuthentication: {}, query parameters: {}", ui.getRequestUri(), qp);
 
 		try {
-			int idpId = Optional.ofNullable((Integer) session.getAttribute(S_IDP_ID))
+			long idpId = Optional.ofNullable((Long) session.getAttribute(S_IDP_ID))
 				.orElseThrow( () -> new AuthenticationException("S_IDP_ID not set"));
 			String token = Optional.ofNullable((String) session.getAttribute(S_CSRF_TOKEN))
 				.orElseThrow( () -> new AuthenticationException("S_CSRF_TOKEN not set") );
